easy_mode: autodetect image format
Autodetecting the disk image format is a controversial topic as, in
general, is considered a dangerous practice. For "easy mode", we're
booting from EDK2 and configuring a single disk image as boot disk.
If we get the image format wrong, EDK2 will fail to find a kernel
and stop there, way before any writes may reach the disk image.

Signed-off-by: Sergio Lopez <slp@redhat.com>