easy_mode: allow using the legecy serial console
Some guests (namely, *BSDs) can't use virtio-console as kernel
console. Since the serial console that's implicitly added for EFI
is not connected to stdio (and we're planning to deprected the EFI
flavor anyway), it can't be used for booting them.

Here we add a command line flag "--serial" that allows users
disabling every console except the legacy serial console.

Signed-off-by: Sergio Lopez <slp@redhat.com>